+++ docs-website/config/navigation.tsx
@@ -417,6 +417,10 @@ const navigation = [
 						title: 'Signing Origin Requests',
 						href: '/docs/guides/signing-origin-requests',
 					},
+					{
+						title: 'Using HTTP proxies',
+						href: '/docs/guides/using-http-proxies',
+					},
 				],
 			},
 		],
@@ -1019,20 +1023,20 @@ const navigation = [
 				href: '/docs/clients-reference/swr',
 			},
 			{
-				title: 'React Query',
-				href: '/docs/clients-reference/react-query',
+				title: 'Next.js',
+				href: '/docs/clients-reference/nextjs',
 			},
 			{
 				title: 'React Relay',
 				href: '/docs/clients-reference/react-relay',
 			},
 			{
-				title: 'Solid Query',
-				href: '/docs/clients-reference/solid-query',
+				title: 'React Query',
+				href: '/docs/clients-reference/react-query',
 			},
 			{
-				title: 'Next.js',
-				href: '/docs/clients-reference/nextjs',
+				title: 'Solid Query',
+				href: '/docs/clients-reference/solid-query',
 			},
 			{
 				title: 'Svelte Query',
@@ -1274,7 +1278,11 @@ const navigation = [
 		title: 'Operations reference',
 		href: '/docs/operations-reference',
 		icon: <CogIcon />,
-		paths: ['/docs/typescript-operations-reference', '/docs/directives-reference'],
+		paths: [
+			'/docs/typescript-operations-reference',
+			'/docs/directives-reference',
+			'/docs/auto-generated-operations-reference',
+		],
 		links: [
 			{
 				title: 'Overview',
@@ -1373,6 +1381,15 @@ const navigation = [
 					},
 				],
 			},
+			{
+				title: 'Auto-generated Operations',
+				links: [
+					{
+						title: 'Overview',
+						href: '/docs/auto-generated-operations-reference',
+					},
+				],
+			},
 		],
 	},
 	{
@@ -1453,6 +1470,10 @@ const navigation = [
 					{
 						title: 'Simplified URL structure',
 						href: '/docs/upgrade-guides/simplifying-the-wundergraph-url-structure',
+					},
+					{
+						title: 'Internal Client deprecated',
+						href: '/docs/upgrade-guides/internal-client-deprecated',
 					},
 				],
 			},