+++ docs-website/src/components/icons/LightbulbIcon.jsx
@@ -1,19 +1,11 @@
-import { DarkMode, Gradient, LightMode } from '@/components/Icon'
+import { DarkMode, Gradient, LightMode } from '@/components/Icon';
 
 export function LightbulbIcon({ id, color }) {
 	return (
 		<>
 			<defs>
-				<Gradient
-					id={`${id}-gradient`}
-					color={color}
-					gradientTransform="matrix(0 21 -21 0 20 11)"
-				/>
-				<Gradient
-					id={`${id}-gradient-dark`}
-					color={color}
-					gradientTransform="matrix(0 24.5001 -19.2498 0 16 5.5)"
-				/>
+				<Gradient id={`${id}-gradient`} color={color} gradientTransform="matrix(0 21 -21 0 20 11)" />
+				<Gradient id={`${id}-gradient-dark`} color={color} gradientTransform="matrix(0 24.5001 -19.2498 0 16 5.5)" />
 			</defs>
 			<LightMode>
 				<circle cx={20} cy={20} r={12} fill={`url(#${id}-gradient)`} />
@@ -42,5 +34,5 @@ export function LightbulbIcon({ id, color }) {
 				/>
 			</DarkMode>
 		</>
-	)
+	);
 }