+++ packages/sdk/src/codegen/index.test.ts
@@ -97,6 +97,13 @@ export const RunTemplateTest = async (...templates: Template[]) => {
 						},
 					},
 					defaultRequestTimeoutSeconds: 0,
+					defaultHttpProxyUrl: {
+						kind: ConfigurationVariableKind.STATIC_CONFIGURATION_VARIABLE,
+						staticVariableContent: '',
+						environmentVariableName: '',
+						environmentVariableDefaultValue: '',
+						placeholderVariableName: '',
+					},
 				},
 				serverOptions: {
 					serverUrl: {
@@ -133,7 +140,7 @@ export const RunTemplateTest = async (...templates: Template[]) => {
 					},
 				},
 				application: {
-					EngineConfiguration: new Api<any>('', [], [], [], []),
+					EngineConfiguration: new Api<any>('', '', [], [], [], []),
 					EnableSingleFlight: true,
 					S3UploadProvider: [],
 					Operations: [
@@ -517,6 +524,7 @@ test('should collect all template dependencies recursively and dedupe based on t
 				},
 			]);
 		}
+
 		dependencies(): Template[] {
 			return [new Template2()];
 		}
@@ -569,6 +577,7 @@ test('should collect templates up to maxTemplateDepth', () => {
 				},
 			]);
 		}
+
 		dependencies(): Template[] {
 			return [new Template2()];
 		}