+++ packages/sdk/src/codegen/templates/typescript/hooks.template.ts
@@ -4,6 +4,7 @@ import { {{ modelImports }} } from "./models"
 import type { BaseRequestContext, WunderGraphRequest, WunderGraphResponse, AuthenticationResponse, AuthenticationHookRequest, HooksConfiguration, WsTransportOnConnectionInitResponse, PreUploadHookRequest, PreUploadHookResponse, PostUploadHookRequest, PostUploadHookResponse } from "@wundergraph/sdk/server";
 import type { InternalClient } from "./wundergraph.internal.client"
 import type { User } from "./wundergraph.server"
+import { InternalOperationsClient } from "./wundergraph.internal.operations.client";
 
 // use SKIP to skip the hook and continue the request / response chain without modifying the request / response
 export type SKIP = "skip";
@@ -34,52 +35,52 @@ export interface WsTransportHookRequest extends BaseRequestContext<User, Interna
 	  dataSourceId: DATA_SOURCES;
 		request: WunderGraphRequest;
 }
-export interface GlobalHooksConfig {
-    httpTransport?: {
-        // onRequest is called right before the request is sent to the origin
-        // it can be used to modify the request
-        // you can return SKIP to skip the hook and continue the request chain without modifying the request
-        // you can return CANCEL to cancel the request chain and return a 500 error
-        onOriginRequest?: {
-            hook: (hook: HttpTransportHookRequest) => Promise<WunderGraphRequest | SKIP | CANCEL>;
-            // calling the httpTransport hooks has a case, because the custom httpTransport hooks have to be called for each request
-            // for this reason, you have to explicitly enable the hook for each Operation
-            enableForOperations?: WUNDERGRAPH_OPERATION[];
-            // enableForAllOperations will disregard the enableForOperations property and enable the hook for all operations
-            enableForAllOperations?: boolean;
-        };
-        // onResponse is called right after the response is received from the origin
-        // it can be used to modify the response
-        // you can return SKIP to skip the hook and continue the response chain without modifying the response
-        // you can return CANCEL to cancel the response chain and return a 500 error
-        onOriginResponse?: {
-            hook: (hook: HttpTransportHookRequestWithResponse) => Promise<WunderGraphResponse | SKIP | CANCEL>;
-            // calling the httpTransport hooks has a case, because the custom httpTransport hooks have to be called for each request
-            // for this reason, you have to explicitly enable the hook for each Operation
-            enableForOperations?: WUNDERGRAPH_OPERATION[];
-            // enableForAllOperations will disregard the enableForOperations property and enable the hook for all operations
-            enableForAllOperations?: boolean;
-        };
-    },
-    wsTransport?: {
-        // onConnectionInit is used to populate 'connection_init' message payload with custom data
-        // it can be used to authenticate the websocket connection
-        onConnectionInit?: {
-            hook: (hook: WsTransportHookRequest) => Promise<WsTransportOnConnectionInitResponse>;
-            /**
-             * enableForDataSources will enable the hook for specific data sources.
-             * you should provide a list of data sources ids
-             * an id is the identifier of the data source in the wundergraph.config.ts file 
-             * @example
-             *const chat = introspect.graphql({             
-             *	id: 'chatId',
-						 *	apiNamespace: 'chat',
-						 *	url: 'http://localhost:8085/query',
-						 *});
-             */
-       	    enableForDataSources: DATA_SOURCES[]; 
-        }
-    }
+export interface GlobalHooksConfig<Operations extends string = string, DataSources extends string = string> {
+	httpTransport?: {
+		// onRequest is called right before the request is sent to the origin
+		// it can be used to modify the request
+		// you can return SKIP to skip the hook and continue the request chain without modifying the request
+		// you can return CANCEL to cancel the request chain and return a 500 error
+		onOriginRequest?: {
+			hook: (hook: HttpTransportHookRequest) => Promise<WunderGraphRequest | SKIP | CANCEL>;
+			// calling the httpTransport hooks has a case, because the custom httpTransport hooks have to be called for each request
+			// for this reason, you have to explicitly enable the hook for each Operation
+			enableForOperations?: Operations[];
+			// enableForAllOperations will disregard the enableForOperations property and enable the hook for all operations
+			enableForAllOperations?: boolean;
+		};
+		// onResponse is called right after the response is received from the origin
+		// it can be used to modify the response
+		// you can return SKIP to skip the hook and continue the response chain without modifying the response
+		// you can return CANCEL to cancel the response chain and return a 500 error
+		onOriginResponse?: {
+			hook: (hook: HttpTransportHookRequestWithResponse) => Promise<WunderGraphResponse | SKIP | CANCEL>;
+			// calling the httpTransport hooks has a case, because the custom httpTransport hooks have to be called for each request
+			// for this reason, you have to explicitly enable the hook for each Operation
+			enableForOperations?: Operations[];
+			// enableForAllOperations will disregard the enableForOperations property and enable the hook for all operations
+			enableForAllOperations?: boolean;
+		};
+	};
+	wsTransport?: {
+		// onConnectionInit is used to populate 'connection_init' message payload with custom data
+		// it can be used to authenticate the websocket connection
+		onConnectionInit?: {
+			hook: (hook: WsTransportHookRequest) => Promise<WsTransportOnConnectionInitResponse>;
+			/**
+			 * enableForDataSources will enable the hook for specific data sources.
+			 * you should provide a list of data sources ids
+			 * an id is the identifier of the data source in the wundergraph.config.ts file
+			 * @example
+			 *const chat = introspect.graphql({
+			 *	id: 'chatId',
+			 *	apiNamespace: 'chat',
+			 *	url: 'http://localhost:8085/query',
+			 *});
+			 */
+			enableForDataSources: DataSources[];
+		};
+	};
 }
         
 export type JSONValue =
@@ -91,7 +92,7 @@ export type JSONValue =
 
 export type JSONObject = { [key: string]: JSONValue };
 										
-export interface HookRequest extends BaseRequestContext<User, InternalClient> {}
+export interface HookRequest extends BaseRequestContext<User, InternalClient, InternalOperationsClient> {}
 
 export interface HookRequestWithResponse<Response> extends HookRequest {
 		response: Response;
@@ -101,26 +102,19 @@ export interface HookRequestWithInput<Input> extends HookRequest {
 		input: Input; 
 }
 						
-export interface HooksConfig extends HooksConfiguration<Queries, Mutations, Subscriptions, Uploads, User, InternalClient> {
-    global?: GlobalHooksConfig;
-    authentication?: {
-        postAuthentication?: (hook: AuthenticationHookRequest<User, InternalClient>) => Promise<void>;
-        mutatingPostAuthentication?: (hook: AuthenticationHookRequest<User, InternalClient>) => Promise<AuthenticationResponse<User>>;
-        revalidate?: (hook: AuthenticationHookRequest<User, InternalClient>) => Promise<AuthenticationResponse<User>>;
-        postLogout?: (hook: AuthenticationHookRequest<User, InternalClient>) => Promise<void>;
-    };
-		{{#if hasQueries}}
-				queries?: Queries;
-		{{/if}}
-		{{#if hasMutations}}
-				mutations?: Mutations;
-		{{/if}}
-		{{#if hasSubscriptions}}
-				subscriptions?: Subscriptions;
-		{{/if}}
-        {{#if hasUploadProviders}}
-                uploads?: Uploads;
-        {{/if}}
+export interface HooksConfig
+	extends HooksConfiguration<Queries, Mutations, Subscriptions, Uploads, User, InternalClient, GlobalHooksConfig<WUNDERGRAPH_OPERATION, DATA_SOURCES>> {
+	authentication?: {
+		postAuthentication?: (hook: AuthenticationHookRequest<User, InternalClient>) => Promise<void>;
+		mutatingPostAuthentication?: (
+			hook: AuthenticationHookRequest<User, InternalClient>
+		) => Promise<AuthenticationResponse<User>>;
+		revalidate?: (hook: AuthenticationHookRequest<User, InternalClient>) => Promise<AuthenticationResponse<User>>;
+		postLogout?: (hook: AuthenticationHookRequest<User, InternalClient>) => Promise<void>;
+	};
+	queries?: Queries;
+	mutations?: Mutations;
+	subscriptions?: Subscriptions;
 }
 
 export interface Queries {