+++ testapps/default/.wundergraph/wundergraph.config.ts
@@ -8,6 +8,7 @@ import {
 } from '@wundergraph/sdk';
 import server from './wundergraph.server';
 import operations from './wundergraph.operations';
+import generate from './wundergraph.generate';
 import { golangClient } from '@wundergraph/golang-client';
 
 const jsp = introspect.openApiV2({
@@ -62,6 +63,7 @@ const countries = introspect.graphql({
 });
 
 const weather = introspect.graphql({
+	id: 'weather',
 	apiNamespace: 'weather',
 	url: 'https://weather-api.wundergraph.com/',
 	introspection: {
@@ -92,25 +94,10 @@ configureWunderGraphApplication({
 	apis: [jsp, weather, countries, spacex, chinook, db, jsp2, usersPost, federatedApi],
 	server,
 	operations,
+	generate,
 	authorization: {
 		roles: ['admin', 'user'],
 	},
-	codeGenerators: [
-		{
-			templates: [
-				// use all the typescript react templates to generate a client
-				...templates.typescript.all,
-			],
-		},
-		{
-			templates: [
-				...golangClient.all({
-					packageName: 'client',
-				}),
-			],
-			path: './generated/golang/client',
-		},
-	],
 	cors: {
 		...cors.allowAll,
 		allowedOrigins: