+++ packages/nextjs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.13](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.10.12...@wundergraph/nextjs@0.10.13) (2023-04-25)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
+## [0.10.12](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.10.11...@wundergraph/nextjs@0.10.12) (2023-04-24)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
 ## [0.10.11](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.10.10...@wundergraph/nextjs@0.10.11) (2023-04-20)
 
 **Note:** Version bump only for package @wundergraph/nextjs